AddNewProduct command was added but not working properly

+++ CoCowork.DataLayer/Entities/Product.cs
@@ -11,7 +11,7 @@ namespace CoCowork.DataLayer.Entities
         public int Id { get; }
         public string Name { get; set; }
         public int Amount { get; set; }
-        public int PriceForOne { get; set; }
+        public decimal PriceForOne { get; set; }
         public string Description { get; set; }
     }
 }
